Changing how we find the default.nix from the jobset.
We were pulling in src as <repo/master>, and using that for the jobset
definition. Instead we can just reuse the hydra configuration's
settings.

For almost all cases, this will be identical. But if you set up a new
hydra build pointing to <mantis-iele-ops/something-other-than-master>,
it will just work. This makes testing new build system change easier.